pyproject.toml
[tool.poetry]
name = "rigel"
version = "0.2.22"
description = "Containerize and deploy your ROS application using Docker."
homepage = "https://github.com/rigel-ros/rigel"
keywords = ["docker", "ros", "cloud", "robotics"] 
authors = ["Pedro Miguel Melo <pedro.m.melo@inesctec.pt>"]
maintainers = ["Pedro Miguel Melo <pedro.m.melo@inesctec.pt>"]
readme = "README.md"
repository = "https://github.com/rigel-ros/rigel"
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
click = "^8.0.3"
Jinja2 = "^3.0.3"
pydantic = "^1.9.0"
python = "^3.8"
python-on-whales = "^0.52.0"
PyYAML = "^6.0"
rich = "^11.2.0"
rigel-hpl = "^0.1.0"
rigel-local-simulation-plugin = "^0.1.0"
rigel-registry-plugin = "^0.1.6"
roslibpy = "^1.3.0"

[tool.poetry.dev-dependencies]
coverage = {extras = ["toml"], version = "^6.3.1"}
flake8 = "^4.0.1"
flake8-bugbear = "^22.1.11"
mypy = "^0.931"
nox = "^2022.1.7"
pytest = "^7.0.1"
pre-commit = "^2.17.0"
twine = "^3.8.0"
types-click = "^7.1.8"
types-pkg-resources = "^0.1.3"
types-PyYAML = "^6.0.4"

[tool.poetry.scripts]
rigel = 'rigel.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
fail_under = 90
